refactor: change add modal naming
